refactor: dedupe context init and locale hash lookup in messages route tests
